cleaned up some warnings with casts and rewrote byte swapping macros not to suck so bad

+++ trunk/MicropolisCore/src/MicropolisEngine/simulate.cpp
@@ -544,7 +544,7 @@ void Micropolis::SetValves()
 
   NormResPop = (float)ResPop / (float)ResPopDenom;
   LastTotalPop = TotalPop;
-  TotalPop = NormResPop + ComPop + IndPop;
+  TotalPop = (short)(NormResPop + ComPop + IndPop);
 
   if (ResPop > 0) {
     Employment = (ComHis[1] + IndHis[1]) / NormResPop;
